modules renaming
