+++ .gitignore
@@ -108,3 +108,4 @@ venv.bak/
 mappings
 aligned.sorted.sam2lca.*
 sam2lca_test_dbdir
+sam2lca.egg-info
