+++ sam2lca/main.py
@@ -1,9 +1,9 @@
-from sam2lca.sam_pysam import Alignment
-from sam2lca.acc2tax_rocksdb import get_mapping
-from sam2lca.acc2tax import get_map_config, base_map_config
+from sam2lca.parse_alignment import Alignment
+from sam2lca.acc2tax import get_mapping
+from sam2lca.data import get_map_config, acc2tax_default
 from sam2lca.lca import compute_lca
 from sam2lca import utils
-from sam2lca.taxonomy_db import setup_taxopy_db, load_taxonomy_db
+from sam2lca.taxonomy import setup_taxopy_db, load_taxonomy_db
 from sam2lca.write_alignment import write_bam_tags
 from pathlib import Path
 import logging
@@ -55,7 +55,7 @@ def sam2lca(
         logging.info(
             f"{acc2tax} acc2tax database is ls not installed. I'm trying to create it for you"
         )
-        get_mapping(map_config=base_map_config, maptype=acc2tax, dbdir=dbdir)
+        get_mapping(map_config=acc2tax_default, maptype=acc2tax, dbdir=dbdir)
     logging.info(f"Step 1/{nb_steps}: Loading taxonomy database")
     TAXDB = load_taxonomy_db(dbdir, taxonomy)
 
@@ -125,11 +125,11 @@ def update_database(
     """
 
     if acc2tax_json is None and acc2tax is not None:
-        map_config = base_map_config
+        map_config = acc2tax_default
     elif acc2tax == "json":
         map_config, acc2tax = get_map_config(map_config_file=map_config)
     if acc2tax is not None:
-        logging.info("* Downloading/updating acc2tax {acc2tax} database ")
+        logging.info(f"* Downloading/updating acc2tax {acc2tax} database ")
         get_mapping(map_config=map_config, maptype=acc2tax, dbdir=dbdir)
 
     if taxonomy:
